Extract runner to new file
